fix: make multisig expenditureFunding properly use multiple tokens
