+++ src/utils/parseFunction.ts
@@ -1,5 +1,6 @@
 import { utils } from 'ethers';
 import { TransactionDescription } from 'ethers/lib/utils';
+import { output } from './logger';
 
 /**
  * Helper attempting to decode function data by trying to parse it with different contract ABIs
@@ -19,4 +20,27 @@ export const parseFunctionData = (
     }
   }
   return null;
+};
+
+/**
+ * @NOTE: This is a rather rudimentary way of handling multicall motions
+ * which only works for multicalls created by UI sagas.
+ * It should be refactored as part of https://github.com/JoinColony/colonyCDapp/issues/2317
+ */
+export const decodeFunctions = (
+  encodedFunctions: string[],
+  interfaces: utils.Interface[],
+): TransactionDescription[] => {
+  const decodedFunctions: TransactionDescription[] = [];
+  for (const functionCall of encodedFunctions) {
+    const parsedFunction = parseFunctionData(functionCall, interfaces);
+    if (!parsedFunction) {
+      output(`Failed to parse multicall function: ${functionCall}`);
+      continue;
+    }
+
+    decodedFunctions.push(parsedFunction);
+  }
+
+  return decodedFunctions;
 };
